fix: #338 联动Sofill=主题，提供统一的hook入口，除了js片段之外，新增挂件触发插槽显示
