+++ package.json
@@ -55,8 +55,8 @@
     "vue-i18n": "^9.2.2"
   },
   "devDependencies": {
-    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
-    "@esbuild-plugins/node-modules-polyfill": "^0.2.0",
+    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
+    "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
     "@types/express": "^4.17.13",
     "@types/node": "^16.18.11",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
